Movieliste aus Datenbank wird angezeigt
