+++ src/components/GenerateMovieList.js
@@ -4,31 +4,14 @@ import { getFirebaseCollectionFrom } from "../firebase";
 import Movielist from "./Movielist";
 
 export default function GenerateMovieList(props) {
-  const [movielist, setMovielist] = useState([]);
-  const getMovieList = () => {
-    getFirebaseCollectionFrom(props.database).onSnapshot((snapshot) => {
-      const dbdata = [];
-      let activeGame = [];
-      snapshot.forEach((doc) => {
-        const data = doc.data();
-        dbdata.push(data);
-        activeGame = dbdata.find((item) => item.active === true);
-      });
-      setMovielist(activeGame.movielist);
-    });
-  };
-
-  useEffect(() => {
-    getMovieList();
-  }, []);
   return (
     <MenuWrapper>
       <Movielist
         thememode={props.thememode}
         apptheme={props.apptheme}
-        movielist={movielist}
+        gamename="Rausvoten"
+        database="rausvoten"
       ></Movielist>
-      {console.log(movielist)}
     </MenuWrapper>
   );
 }
